Badge reorganization (#1505)
* moving Badge props ts to types.ts

* changeset

* extending darkModeProps

* adding Badge subdirectory
