+++ .changeset/pretty-apricots-develop.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/badge': patch
+---
+
+Standardizing file structure
