+++ packages/badge/src/Badge.story.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, Meta } from '@storybook/react';
 import React from 'react';
-import Badge from './Badge';
+import Badge from './Badge/Badge';
 import { storybookArgTypes } from '@leafygreen-ui/lib';
 
 export default {
