--- packages/badge/src/types.ts
@@ -1,10 +1,0 @@
-export const Variant = {
-  DarkGray: 'darkgray',
-  LightGray: 'lightgray',
-  Red: 'red',
-  Yellow: 'yellow',
-  Blue: 'blue',
-  Green: 'green',
-} as const;
-
-export type Variant = typeof Variant[keyof typeof Variant];
